pt_decomposers -> decomposers
